refactor(hf): split backend helper modules
